二分查找，441，排列硬币，二分查找（还提交了个错的。。。醉了）

+++ binary_search/arranging_coins/ArrangingCoins.java
@@ -2,18 +2,18 @@ package binary_search.arranging_coins;
 
 public class ArrangingCoins {
     public int arrangeCoins(int n) {
-        int lo = 1, hi = n;
+        long lo = 1, hi = n;
         while (lo <= hi) {
-            int mid = lo + (hi - lo) / 2;
+            long mid = lo + (hi - lo) / 2;
             long sum = (mid + 1) * mid / 2;
             if (sum == n) {
-                return mid;
+                return (int)mid;
             } else if (sum > n) {
                 hi = mid - 1;
             } else {
                 lo = mid + 1;
             }
         }
-        return lo - 1;
+        return (int)lo - 1;
     }
 }
